# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Virtual Shelf Book Service API
  version: 0.10.0
  description: |
    Virtual Shelf MSA의 Book Service 계약이다.
    알라딘 API 기반 외부 도서 검색, 개인 서재 관리, 독서 진도, 문장 스크랩, 동물 사서를 제공한다.

    표지 OCR과 AI 기반 도서/무드 분석은 다른 MSA 컴포넌트가 담당하며 이 계약의 범위 밖이다.
    이 API Server는 이미 인식된 isbn을 넘겨받아 외부 도서 API 조회, 독서 진도 계산,
    서재·스크랩·동물 사서 CRUD 등 단순 데이터베이스 CRUD로 해결 가능한 부분만 다룬다.

    모든 API는 인증이 필요하다. 사용자 식별자는 요청 body가 아니라 검증된 인증 principal에서 얻는다.
  contact:
    name: Virtual Shelf Book Service Team
servers:
  - url: /
    description: 현재 호스트
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/librarian-types:
    get:
      tags: [Librarian]
      operationId: getLibrarianTypes
      summary: 사서 타입 카탈로그 조회
      description: 획득 가능한 동물 사서 타입(러시안블루, 슈빌)별 기본/클릭 시 이미지를 조회한다.
      responses:
        "200":
          description: 사서 타입 카탈로그 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibrarianTypeListResponse"
              example:
                types:
                  - type: RUSSIAN_BLUE
                    imageUrl: https://example.com/librarians/cat-1.png
                    clickedImageUrl: https://example.com/librarians/cat-1-clicked.png
                  - type: SHOEBILL
                    imageUrl: https://example.com/librarians/bird-1.png
                    clickedImageUrl: https://example.com/librarians/bird-1-clicked.png
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/librarians:
    post:
      tags: [Librarian]
      operationId: acquireLibrarian
      summary: 사서 획득
      description: |
        로그인 사용자가 지정한 타입의 사서를 새로 획득해 이름을 지어준다.
        같은 타입의 사서는 회원당 최대 1마리만 보유할 수 있다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcquireLibrarianRequest"
            example:
              type: RUSSIAN_BLUE
              name: 나비
      responses:
        "201":
          description: 사서 획득 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcquireLibrarianResponse"
              example:
                librarianId: 10
                type: RUSSIAN_BLUE
                name: 나비
                level: 1
                experience: 0
                isRepresentative: false
                createdAt: "2026-08-25T10:00:00+09:00"
        "400":
          $ref: "#/components/responses/InvalidLibrarianData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/LibrarianAlreadyOwned"

    get:
      tags: [Librarian]
      operationId: getLibrarians
      summary: 내가 보유한 사서 목록 조회
      description: 로그인 사용자가 보유한 사서 전체를 조회한다.
      responses:
        "200":
          description: 보유 사서 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibrarianListResponse"
              example:
                librarians:
                  - librarianId: 10
                    type: RUSSIAN_BLUE
                    name: 나비
                    level: 1
                    experience: 0
                    isRepresentative: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/librarians/representative:
    get:
      tags: [Librarian]
      operationId: getRepresentative
      summary: 대표 사서 조회
      description: 로그인 사용자가 지정한 대표 사서를 조회한다. 아직 지정하지 않았으면 404를 반환한다.
      responses:
        "200":
          description: 대표 사서 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepresentativeLibrarianResponse"
              example:
                librarianId: 10
                type: RUSSIAN_BLUE
                name: 나비
                level: 1
                updatedAt: "2026-08-25T10:00:00+09:00"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/RepresentativeLibrarianNotSelected"
  /api/v1/librarians/{librarianId}:
    parameters:
      - $ref: "#/components/parameters/LibrarianId"
    patch:
      tags: [Librarian]
      operationId: renameLibrarian
      summary: 사서 이름 변경
      description: 로그인 사용자가 보유한 사서의 이름을 바꾼다. 언제든 몇 번이든 바꿀 수 있다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameLibrarianRequest"
            example:
              name: 루루
      responses:
        "200":
          description: 사서 이름 변경 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenameLibrarianResponse"
              example:
                librarianId: 10
                name: 루루
                updatedAt: "2026-08-25T10:05:00+09:00"
        "400":
          $ref: "#/components/responses/InvalidLibrarianData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/LibrarianAccessDenied"
        "404":
          $ref: "#/components/responses/LibrarianNotFound"

    delete:
      tags: [Librarian]
      operationId: deleteLibrarian
      summary: 사서 방출
      description: 로그인 사용자가 보유한 사서를 방출(논리 삭제)한다.
      responses:
        "204":
          description: 사서 방출 성공. 응답 body 없음
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/LibrarianAccessDenied"
        "404":
          $ref: "#/components/responses/LibrarianNotFound"
  /api/v1/librarians/{librarianId}/representative:
    parameters:
      - $ref: "#/components/parameters/LibrarianId"
    patch:
      tags: [Librarian]
      operationId: selectRepresentative
      summary: 대표 사서 지정
      description: |
        로그인 사용자가 보유한 사서 중 하나를 대표 사서로 지정한다. 이미 다른 사서가 대표로 지정되어 있으면
        그 사서는 자동으로 대표에서 해제된다. 회원당 대표 사서는 항상 최대 1마리다.
      responses:
        "200":
          description: 대표 사서 지정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepresentativeLibrarianResponse"
              example:
                librarianId: 10
                type: RUSSIAN_BLUE
                name: 나비
                level: 1
                updatedAt: "2026-08-25T10:10:00+09:00"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/LibrarianAccessDenied"
        "404":
          $ref: "#/components/responses/LibrarianNotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 인증 서비스가 발급하고 Book Service가 검증하는 access token
  parameters:
...
    LibrarianId:
      name: librarianId
      in: path
      required: true
      description: 회원이 보유한 사서 식별자
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 10
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          minLength: 1
          description: 클라이언트 분기용 안정적인 오류 코드
        message:
          type: string
          minLength: 1
          description: 사용자에게 표시 가능한 한국어 오류 메시지
...
    LibrarianType:
      type: string
      enum: [RUSSIAN_BLUE, SHOEBILL]
      description: 획득 가능한 사서 종류. 타입 자체가 늘어나면 계약도 함께 갱신한다.
      example: RUSSIAN_BLUE
    LibrarianTypeSummary:
      type: object
      additionalProperties: false
      required: [type, imageUrl, clickedImageUrl]
      properties:
        type:
          $ref: "#/components/schemas/LibrarianType"
        imageUrl:
          type: string
          format: uri
        clickedImageUrl:
          type: string
          format: uri
    LibrarianTypeListResponse:
      type: object
      additionalProperties: false
      required: [types]
      properties:
        types:
          type: array
          items:
            $ref: "#/components/schemas/LibrarianTypeSummary"
    AcquireLibrarianRequest:
      type: object
      required: [type, name]
      properties:
        type:
          $ref: "#/components/schemas/LibrarianType"
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: 회원이 직접 짓는 사서 이름. 서버가 타입 이름을 복사해 채우지 않는다.
    AcquireLibrarianResponse:
      type: object
      additionalProperties: false
      required: [librarianId, type, name, level, experience, isRepresentative, createdAt]
      properties:
        librarianId:
          type: integer
          format: int64
          minimum: 1
        type:
          $ref: "#/components/schemas/LibrarianType"
        name:
          type: string
        level:
          type: integer
          format: int32
          minimum: 1
        experience:
          type: integer
          format: int64
          minimum: 0
        isRepresentative:
          type: boolean
        createdAt:
          type: string
          format: date-time
    LibrarianSummary:
      type: object
      additionalProperties: false
      required: [librarianId, type, name, level, experience, isRepresentative]
      properties:
        librarianId:
          type: integer
          format: int64
          minimum: 1
        type:
          $ref: "#/components/schemas/LibrarianType"
        name:
          type: string
        level:
          type: integer
          format: int32
          minimum: 1
        experience:
          type: integer
          format: int64
          minimum: 0
        isRepresentative:
          type: boolean
    LibrarianListResponse:
      type: object
      additionalProperties: false
      required: [librarians]
      properties:
        librarians:
          type: array
          items:
            $ref: "#/components/schemas/LibrarianSummary"
    RenameLibrarianRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
    RenameLibrarianResponse:
      type: object
      additionalProperties: false
      required: [librarianId, name, updatedAt]
      properties:
        librarianId:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        updatedAt:
          type: string
          format: date-time
    RepresentativeLibrarianResponse:
      type: object
      additionalProperties: false
      required: [librarianId, type, name, level, updatedAt]
      properties:
        librarianId:
          type: integer
          format: int64
          minimum: 1
        type:
          $ref: "#/components/schemas/LibrarianType"
        name:
          type: string
        level:
          type: integer
          format: int32
          minimum: 1
        updatedAt:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: UNAUTHORIZED
            message: 인증이 필요합니다.
...
    InvalidLibrarianData:
      description: 올바르지 않은 사서 정보(이름 공백 등)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INVALID_LIBRARIAN_DATA
            message: 올바르지 않은 사서 정보입니다.
    LibrarianAlreadyOwned:
      description: 같은 타입의 사서를 이미 보유함
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARIAN_ALREADY_OWNED
            message: 이미 해당 타입의 사서를 보유하고 있습니다.
    LibrarianAccessDenied:
      description: 사서 접근 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARIAN_ACCESS_DENIED
            message: 해당 사서에 접근할 권한이 없습니다.
    LibrarianNotFound:
      description: 사서 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARIAN_NOT_FOUND
            message: 사서를 찾을 수 없습니다.
    RepresentativeLibrarianNotSelected:
      description: 대표 사서를 아직 선택하지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: REPRESENTATIVE_LIBRARIAN_NOT_SELECTED
            message: 대표 사서를 아직 선택하지 않았습니다.